@@ -18,88 +35,96 @@
     "lint": "npm run lint:fix && npm run lint:style",
     "lint:ts": "ng lint",
     "lint:fix": "ng lint --fix",
-    "lint:style": "stylelint \"src/**/*.less\" --syntax less --fix"
+    "lint:style": "stylelint \"src/**/*.less\" --syntax less --fix",
+    "package": "ng build --configuration production"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.2.0",
-    "@angular/cdk": "~12.2.0",
-    "@angular/common": "~12.2.0",
-    "@angular/compiler": "~12.2.0",
-    "@angular/core": "~12.2.0",
-    "@angular/forms": "~12.2.0",
-    "@angular/platform-browser": "~12.2.0",
-    "@angular/platform-browser-dynamic": "~12.2.0",
-    "@angular/router": "~12.2.0",
-    "@delon/abc": "^12.4.2",
-    "@delon/acl": "^12.4.2",
-    "@delon/auth": "^12.4.2",
-    "@delon/cache": "^12.4.2",
-    "@delon/chart": "^12.4.2",
-    "@delon/form": "^12.4.2",
-    "@delon/mock": "^12.4.2",
-    "@delon/theme": "^12.4.2",
-    "@delon/util": "^12.4.2",
-    "ajv": "^8.6.2",
-    "ajv-formats": "^2.1.1",
-    "echarts": "^5.2.2",
-    "ng-alain": "^12.4.2",
-    "ng-zorro-antd": "^12.0.2",
-    "ngx-color-picker": "^12.0.1",
-    "ngx-echarts": "^v7.1.0",
-    "rxjs": "~6.6.0",
-    "screenfull": "^5.1.0",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "@angular/animations": "15.2.10",
+    "@angular/common": "15.2.10",
+    "@angular/compiler": "15.2.10",
+    "@angular/core": "15.2.10",
+    "@angular/forms": "15.2.10",
+    "@angular/platform-browser": "15.2.10",
+    "@angular/platform-browser-dynamic": "15.2.10",
+    "@angular/router": "15.2.10",
+    "@delon/abc": "15.2.1",
+    "@delon/acl": "15.2.1",
+    "@delon/auth": "15.2.1",
+    "@delon/cache": "15.2.1",
+    "@delon/chart": "15.2.1",
+    "@delon/form": "15.2.1",
+    "@delon/mock": "15.2.1",
+    "@delon/theme": "15.2.1",
+    "@delon/util": "15.2.1",
+    "ajv": "8.12.0",
+    "ajv-formats": "2.1.1",
+    "angular-tag-cloud-module": "15.0.0",
+    "echarts": "5.4.3",
+    "jquery": "3.7.1",
+    "monaco-editor": "0.36.1",
+    "ng-zorro-antd": "15.1.0",
+    "ngx-color-picker": "12.0.1",
+    "ngx-echarts": "15.0.3",
+    "ngx-slick-carousel": "15.0.0",
+    "rxjs": "7.8.1",
+    "screenfull": "6.0.2",
+    "slick-carousel": "1.8.1",
+    "tslib": "2.6.2",
+    "zone.js": "0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.2.13",
-    "@angular-eslint/builder": "~12.3.1",
-    "@angular-eslint/eslint-plugin": "~12.3.1",
-    "@angular-eslint/eslint-plugin-template": "~12.3.1",
-    "@angular-eslint/schematics": "~12.3.1",
-    "@angular-eslint/template-parser": "~12.3.1",
-    "@angular/cli": "~12.2.13",
-    "@angular/compiler-cli": "~12.2.0",
-    "@angular/language-service": "~12.2.0",
-    "@delon/testing": "^12.4.2",
-    "@ngx-formly/schematics": "^5.10.23",
-    "@types/echarts": "^4.9.12",
-    "@types/jasmine": "~3.8.0",
-    "@types/node": "^12.11.1",
-    "@typescript-eslint/eslint-plugin": "~4.29.2",
-    "@typescript-eslint/parser": "~4.29.2",
-    "eslint": "^7.32.0",
-    "eslint-config-prettier": "^2.2.1",
-    "eslint-plugin-import": "~2.24.1",
-    "eslint-plugin-jsdoc": "~36.0.7",
-    "eslint-plugin-prefer-arrow": "~1.2.3",
-    "eslint-plugin-prettier": "^2.2.1",
-    "jasmine-core": "~3.8.0",
-    "karma": "~6.3.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "~1.7.0",
-    "lint-staged": "^11.1.2",
-    "ng-alain": "^12.4.2",
-    "ng-alain-plugin-theme": "^12.0.0",
-    "prettier": "^2.2.1",
-    "source-map-explorer": "^2.5.2",
-    "stylelint": "^13.13.1",
-    "stylelint-config-prettier": "^8.0.2",
-    "stylelint-config-rational-order": "^0.1.2",
-    "stylelint-config-standard": "^22.0.0",
-    "stylelint-declaration-block-no-ignored-properties": "^2.4.0",
-    "stylelint-order": "^4.1.0",
-    "typescript": "~4.3.5"
+    "@angular-devkit/build-angular": "15.2.10",
+    "@angular-eslint/builder": "15.2.1",
+    "@angular-eslint/eslint-plugin": "15.2.1",
+    "@angular-eslint/eslint-plugin-template": "15.2.1",
+    "@angular-eslint/schematics": "15.2.1",
+    "@angular-eslint/template-parser": "15.2.1",
+    "@angular/cli": "15.2.10",
+    "@angular/compiler-cli": "15.2.10",
+    "@angular/language-service": "15.2.10",
+    "@delon/testing": "15.2.1",
+    "@types/jasmine": "4.0.3",
+    "@types/jasminewd2": "2.0.12",
+    "@types/node": "12.20.55",
+    "@typescript-eslint/eslint-plugin": "5.49.0",
+    "@typescript-eslint/parser": "5.49.0",
+    "eslint": "8.51.0",
+    "eslint-config-prettier": "8.6.0",
+    "eslint-plugin-deprecation": "1.3.3",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jsdoc": "39.7.5",
+    "eslint-plugin-prefer-arrow": "1.2.3",
+    "eslint-plugin-prettier": "4.2.1",
+    "husky": "7.0.4",
+    "jasmine-core": "4.3.0",
+    "jasmine-spec-reporter": "7.0.0",
+    "karma": "6.4.2",
+    "karma-chrome-launcher": "3.1.1",
+    "karma-coverage": "2.2.1",
+    "karma-jasmine": "5.1.0",
+    "karma-jasmine-html-reporter": "2.0.0",
+    "lint-staged": "13.3.0",
+    "ng-alain": "15.2.1",
+    "node-fetch": "2.7.0",
+    "prettier": "2.8.8",
+    "protractor": "7.0.0",
+    "source-map-explorer": "2.5.3",
+    "stylelint": "14.16.1",
+    "stylelint-config-prettier": "9.0.5",
+    "stylelint-config-rational-order": "0.1.2",
+    "stylelint-config-standard": "28.0.0",
+    "stylelint-declaration-block-no-ignored-properties": "2.7.0",
+    "stylelint-order": "5.0.0",
+    "ts-node": "10.9.1",
+    "typescript": "4.9.5"
   },
   "lint-staged": {
     "(src)/**/*.{html,ts}": [
       "eslint --fix"
     ],
     "(src)/**/*.less": [
-      "stylelint --syntax less --fix"
+      "npm run lint:style"
     ]
   }
 }
