@@ -20,196 +20,146 @@ import cdnTransfer from '../CdnTransfer'
 import styles from './styles.module.css'
 import {features, dromaraFriends, friendsLink, usersLink} from '../constants'
 import LogoCarouselStatic from './components/LogoCarouselStatic'
+import Sponsor from "./components/Sponsor";
+import DiaLog from "./components/DiaLog";
 
 function Home() {
   const context = useDocusaurusContext()
   useEffect(() => autoRedirect(), [])
   const { siteConfig = {} } = context
   return (
-    <Layout
-      title={`${siteConfig.title} · ${siteConfig.tagline}`}
-      description={`${siteConfig.tagline}`}
-    >
-      <header className={clsx('hero hero--primary', styles.heroBanner)}>
-        <div className="container">
-          <h1 className="hero__title">
-            <img
-              style={{ width: '500px', marginTop: '100px' }}
-              src={cdnTransfer('/img/hertzbeat-brand.svg')}
-              alt={'#'}
+      <>
+        <Layout
+            title={`${siteConfig.title} · ${siteConfig.tagline}`}
+            description={`${siteConfig.tagline}`}
+        >
+          <header className={clsx('hero hero--primary', styles.heroBanner)}>
+            <div className="container">
+              <h1 className="hero__title">
+                <img
+                    style={{ width: '500px', marginTop: '100px' }}
+                    src={cdnTransfer('/img/hertzbeat-brand.svg')}
+                    alt={'#'}
+                />
+              </h1>
+              <p className="hero__subtitle">
+                <Translate>slogan</Translate>
+              </p>
+              <div className={styles.buttons}>
+                <Link
+                    to="/docs/"
+                    className={clsx(
+                        'button button--primary button--lg',
+                        styles.getStarted
+                    )}
+                >
+                  <Translate>quickstart</Translate>
+                </Link>
+              </div>
+            </div>
+          </header>
+          <main>
+            <div className={styles.sectionDashboard}>
+              <Swiper
+                  modules={[Autoplay, EffectFade, Navigation]}
+                  watchSlidesProgress={true}
+                  navigation={{
+                    nextEl: '.user-swiper-button-next',
+                    prevEl: '.user-swiper-button-prev',
+                  }}
+                  grabCursor
+                  // 轮播下用这个效果会失效
+                  // effect={'fade'}
+                  // fadeEffect={{
+                  //   crossFade: true
+                  // }}
+                  // slidesPerView={1}
+                  // 自动轮播
+                  loop={true}
+                  speed={0}
+                  autoplay={{
+                    delay: 6000,
+                    disableOnInteraction: false,
+                    waitForTransition: false,
+                  }}
+              >
+                <SwiperSlide>
+                  <img
+                      style={{ width: '1400px', display: 'block', margin: '0 auto' }}
+                      src={useBaseUrl('/img/docs/hertzbeat-arch.png')}
+                  />
+                </SwiperSlide>
+                <SwiperSlide>
+                  <img
+                      style={{ width: '1400px', display: 'block', margin: '0 auto' }}
+                      src={useBaseUrl('/img/home/status.png')}
+                  />
+                </SwiperSlide>
+                <SwiperSlide>
+                  <img
+                      style={{ width: '1400px', display: 'block', margin: '0 auto' }}
+                      src={useBaseUrl('/img/home/0.png')}
+                  />
+                </SwiperSlide>
+                <SwiperSlide>
+                  <img
+                      style={{ width: '1400px', display: 'block', margin: '0 auto' }}
+                      src={useBaseUrl('/img/home/1.png')}
+                  />
+                </SwiperSlide>
+                <SwiperSlide>
+                  <img
+                      style={{ width: '1400px', display: 'block', margin: '0 auto' }}
+                      src={useBaseUrl('/img/home/9.png')}
+                  />
+                </SwiperSlide>
+              </Swiper>
+            </div>
+
+            <div
+                className="swiper-button-prev user-swiper-button-prev"
+                style={{ top: '880px', left: '50px', color: '#000033' }}
             />
-          </h1>
-          <p className="hero__subtitle">
-            <Translate>易用友好的开源实时监控系统</Translate>
-          </p>
-          <div className={styles.social}>
-            <a href="https://gitter.im/hertzbeat/community?utm_source=badge&utm_medium=badge&utm_campaign=pr-badge">
-              <img
-                src={'https://badges.gitter.im/hertzbeat/community.svg'}
-                alt={''}
-              />
-            </a>
-            <a href="https://console.tancloud.cn">
-              <img src={cdnTransfer('/img/badge/web-monitor.svg')} alt={''} />
-            </a>
-            <a href="https://console.tancloud.cn">
-              <img src={cdnTransfer('/img/badge/ping-connect.svg')} alt={''} />
-            </a>
-            <a href="https://console.tancloud.cn">
-              <img
-                src={cdnTransfer('/img/badge/port-available.svg')}
-                alt={''}
-              />
-            </a>
-            <a href="https://console.tancloud.cn">
-              <img
-                src={cdnTransfer('/img/badge/database-monitor.svg')}
-                alt={''}
-              />
-            </a>
-            <a href="https://console.tancloud.cn">
-              <img src={cdnTransfer('/img/badge/os-monitor.svg')} alt={''} />
-            </a>
-            <a href="https://console.tancloud.cn">
-              <img
-                src={cdnTransfer('/img/badge/custom-monitor.svg')}
-                alt={''}
-              />
-            </a>
-            <a href="https://console.tancloud.cn">
-              <img src={cdnTransfer('/img/badge/threshold.svg')} alt={''} />
-            </a>
-            <a href="https://console.tancloud.cn">
-              <img src={cdnTransfer('/img/badge/alert.svg')} alt={''} />
-            </a>
-          </div>
-          <div className={styles.buttons}>
-            <Link
-              to="https://console.tancloud.cn"
-              className={clsx(
-                'button button--outline button--secondary button--lg',
-                styles.getStarted
-              )}
-            >
-              <Translate>即刻登录免费使用</Translate>
-            </Link>
-          </div>
-        </div>
-      </header>
-      <main>
-        <div className={styles.sectionDashboard}>
-          <Swiper
-            modules={[Autoplay, EffectFade, Navigation]}
-            watchSlidesProgress={true}
-            navigation={{
-              nextEl: '.user-swiper-button-next',
-              prevEl: '.user-swiper-button-prev',
-            }}
-            grabCursor
-            // 轮播下用这个效果会失效
-            // effect={'fade'}
-            // fadeEffect={{
-            //   crossFade: true
-            // }}
-            // slidesPerView={1}
-            // 自动轮播
-            loop={true}
-            speed={0}
-            autoplay={{
-              delay: 3000,
-              disableOnInteraction: false,
-              waitForTransition: false,
-            }}
-          >
-            <SwiperSlide>
-              <img
-                style={{ width: '1000px', display: 'block', margin: '0 auto' }}
-                src={useBaseUrl('/img/home/1.png')}
-              />
-            </SwiperSlide>
-            <SwiperSlide>
-              <img
-                style={{ width: '1000px', display: 'block', margin: '0 auto' }}
-                src={useBaseUrl('/img/home/2.png')}
-              />
-            </SwiperSlide>
-            <SwiperSlide>
-              <img
-                style={{ width: '1000px', display: 'block', margin: '0 auto' }}
-                src={useBaseUrl('/img/home/3.png')}
-              />
-            </SwiperSlide>
-            <SwiperSlide>
-              <img
-                style={{ width: '1000px', display: 'block', margin: '0 auto' }}
-                src={useBaseUrl('/img/home/4.png')}
-              />
-            </SwiperSlide>
-            <SwiperSlide>
-              <img
-                style={{ width: '1000px', display: 'block', margin: '0 auto' }}
-                src={useBaseUrl('/img/home/5.png')}
-              />
-            </SwiperSlide>
-            <SwiperSlide>
-              <img
-                style={{ width: '1000px', display: 'block', margin: '0 auto' }}
-                src={useBaseUrl('/img/home/6.png')}
-              />
-            </SwiperSlide>
-            <SwiperSlide>
-              <img
-                style={{ width: '1000px', display: 'block', margin: '0 auto' }}
-                src={useBaseUrl('/img/home/7.png')}
-              />
-            </SwiperSlide>
-            <SwiperSlide>
-              <img
-                style={{ width: '1000px', display: 'block', margin: '0 auto' }}
-                src={useBaseUrl('/img/home/8.png')}
-              />
-            </SwiperSlide>
-          </Swiper>
-        </div>
+            <div
+                className="swiper-button-next user-swiper-button-next"
+                style={{ top: '880px', right: '50px', color: '#000033' }}
+            />
+            {/*who is using*/}
+            <Section isDark>
+              <LogoCarousel logos={usersLink}
+                            headerTitle={translate({ message: 'Who uses HertzBeat?' })}>
+              </LogoCarousel>
+            </Section>
+            {features && features.length > 0 && (
+                <Section>
+                  {features.map((props, idx) => (
+                      <Feature key={idx} {...props} />
+                  ))}
+                </Section>
+            )}
 
-        <div
-          className="swiper-button-prev user-swiper-button-prev"
-          style={{ top: '880px', left: '50px', color: '#000033' }}
-        />
-        <div
-          className="swiper-button-next user-swiper-button-next"
-          style={{ top: '880px', right: '50px', color: '#000033' }}
-        />
-        {/*who is using*/}
-        <Section isDark>
-            <LogoCarousel logos={usersLink}
-                          headerTitle={translate({ message: 'Who Use HertzBeat?' })}>
-            </LogoCarousel>
-        </Section>
-        {features && features.length > 0 && (
-          <Section>
-            {features.map((props, idx) => (
-              <Feature key={idx} {...props} />
-            ))}
-          </Section>
-        )}
+            {/*<Section>*/}
+            {/*  <Sponsor headerTitle={translate({ message: 'sponsor' })}/>*/}
+            {/*</Section>*/}
 
-        {/*Friend Links*/}
-        <Section>
-          <LogoCarousel
-            logos={dromaraFriends}
-            headerTitle={translate({ message: 'Dromara Friends' })}
-          />
-        </Section>
-        {/*/!*Friend Links*!/*/}
-        <Section>
-          <LogoCarouselStatic
-            logos={friendsLink}
-            headerTitle={translate({ message: 'Friends Link' })}
-          />
-        </Section>
-      </main>
-    </Layout>
+            {/*Friend Links*/}
+            <Section>
+              <LogoCarousel
+                  logos={dromaraFriends}
+                  headerTitle={translate({ message: 'Dromara Friends' })}
+              />
+            </Section>
+            {/*/!*Friend Links*!/*/}
+            {/*<Section>*/}
+            {/*  <LogoCarouselStatic*/}
+            {/*      logos={friendsLink}*/}
+            {/*      headerTitle={translate({ message: 'Friends Link' })}*/}
+            {/*  />*/}
+            {/*</Section>*/}
+          </main>
+        </Layout>
+        <DiaLog/>
+      </>
   )
 }
 
@@ -236,11 +186,12 @@ function loadGitter() {
 function autoRedirect() {
   let lang = global.navigator?.language || navigator?.userLanguage
   console.log('Current lang is ' + lang)
-  if (lang != null && lang.toLowerCase() !== 'zh-cn') {
-    if (sessionStorage.getItem('auto_detect_redirect') !== 'true') {
-      console.log('Not zh-cn, redirect to en')
+  if (lang != null && (lang.toLowerCase() === 'zh-cn' || lang.toLowerCase().indexOf('zh') > 0)) {
+    console.log(window.location.pathname);
+    if (sessionStorage.getItem('auto_detect_redirect') !== 'true' && !window.location.pathname.startsWith('/zh-cn', false)) {
+      console.log('current lang is zh-cn, redirect to zh-cn')
       sessionStorage.setItem('auto_detect_redirect', 'true')
-      window.location.href = '/en'
+      window.location.href = '/zh-cn'
     }
   }
 }
