@@ -1,6 +1,7 @@
-import { Component, ElementRef, OnInit, Renderer2 } from '@angular/core';
+import { Component, ElementRef, Inject, OnInit, Renderer2 } from '@angular/core';
 import { NavigationEnd, NavigationError, RouteConfigLoadStart, Router } from '@angular/router';
-import { TitleService, VERSION as VERSION_ALAIN } from '@delon/theme';
+import { I18NService } from '@core';
+import { ALAIN_I18N_TOKEN, TitleService, VERSION as VERSION_ALAIN } from '@delon/theme';
 import { environment } from '@env/environment';
 import { NzModalService } from 'ng-zorro-antd/modal';
 import { VERSION as VERSION_ZORRO } from 'ng-zorro-antd/version';
@@ -15,7 +16,8 @@ export class AppComponent implements OnInit {
     renderer: Renderer2,
     private router: Router,
     private titleSrv: TitleService,
-    private modalSrv: NzModalService
+    private modalSrv: NzModalService,
+    @Inject(ALAIN_I18N_TOKEN) private i18nSvc: I18NService
   ) {
     renderer.setAttribute(el.nativeElement, 'ng-alain-version', VERSION_ALAIN.full);
     renderer.setAttribute(el.nativeElement, 'ng-zorro-version', VERSION_ZORRO.full);
@@ -29,11 +31,13 @@ export class AppComponent implements OnInit {
       }
       if (configLoad && ev instanceof NavigationError) {
         this.modalSrv.confirm({
-          nzTitle: `提醒`,
-          nzContent: environment.production ? `应用可能已发布新版本，请点击刷新才能生效。` : `无法加载路由：${ev.url}`,
+          nzTitle: this.i18nSvc.fanyi('common.notice'),
+          nzContent: environment.production
+            ? `New version may have been published, please click refresh.`
+            : `Can not resolve route：${ev.url}`,
           nzCancelDisabled: false,
-          nzOkText: '刷新',
-          nzCancelText: '忽略',
+          nzOkText: this.i18nSvc.fanyi('common.refresh'),
+          nzCancelText: this.i18nSvc.fanyi('common.ignore'),
           nzOnOk: () => location.reload()
         });
       }
