@@ -15,7 +15,7 @@ import { DELON_LOCALE, en_US as delonLang } from '@delon/theme';
 import { enUS as dateLang } from 'date-fns/locale';
 import { NZ_DATE_LOCALE, NZ_I18N, en_US as zorroLang } from 'ng-zorro-antd/i18n';
 const LANG = {
-  abbr: 'en',
+  abbr: 'en-US',
   ng: ngLang,
   zorro: zorroLang,
   date: dateLang,
@@ -50,7 +50,7 @@ const INTERCEPTOR_PROVIDES = [{ provide: HTTP_INTERCEPTORS, useClass: DefaultInt
 // #endregion
 
 // #region global third module
-const GLOBAL_THIRD_MODULES: Array<Type<void>> = [];
+const GLOBAL_THIRD_MODULES: Array<Type<void>> = [SlickCarouselModule, TagCloudComponent];
 // #endregion
 
 // #region Startup Service
@@ -78,6 +78,8 @@ import { SharedModule } from './shared/shared.module';
 import { STWidgetModule } from './shared/st-widget/st-widget.module';
 import { ReactiveFormsModule } from '@angular/forms';
 import { NgxEchartsModule } from 'ngx-echarts';
+import { SlickCarouselModule } from 'ngx-slick-carousel';
+import { TagCloudComponent } from 'angular-tag-cloud-module';
 
 @NgModule({
   declarations: [AppComponent],
