@@ -1,5 +1,7 @@
-import { Injectable } from '@angular/core';
+import { Inject, Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, CanActivate, Router, RouterStateSnapshot, UrlTree } from '@angular/router';
+import { I18NService } from '@core';
+import { ALAIN_I18N_TOKEN } from '@delon/theme';
 import { NzNotificationService } from 'ng-zorro-antd/notification';
 import { Observable } from 'rxjs';
 
@@ -9,7 +11,12 @@ import { LocalStorageService } from '../../service/local-storage.service';
   providedIn: 'root'
 })
 export class DetectAuthGuard implements CanActivate {
-  constructor(private localStorageSvc: LocalStorageService, private notifySvc: NzNotificationService, private router: Router) {}
+  constructor(
+    private localStorageSvc: LocalStorageService,
+    private notifySvc: NzNotificationService,
+    @Inject(ALAIN_I18N_TOKEN) private i18nSvc: I18NService,
+    private router: Router
+  ) {}
 
   canActivate(
     route: ActivatedRouteSnapshot,
@@ -18,7 +25,7 @@ export class DetectAuthGuard implements CanActivate {
     let activate = this.localStorageSvc.hasAuthorizationToken();
     if (!activate) {
       setTimeout(() => {
-        this.notifySvc.warning('请先登录!', '');
+        this.notifySvc.warning(this.i18nSvc.fanyi('app.login.notify'), '');
         this.router.navigateByUrl('/passport/login');
       });
     }
