@@ -1,8 +1,10 @@
-import { Component } from '@angular/core';
-import { SettingsService, User } from '@delon/theme';
+import { Component, Inject } from '@angular/core';
+import { I18NService } from '@core';
+import { ALAIN_I18N_TOKEN, SettingsService, User } from '@delon/theme';
 import { LayoutDefaultOptions } from '@delon/theme/layout-default';
 import { environment } from '@env/environment';
-import { CONSTS } from 'src/app/shared/consts';
+
+import { CONSTANTS } from '../../shared/constants';
 
 @Component({
   selector: 'layout-basic',
@@ -48,7 +50,7 @@ import { CONSTS } from 'src/app/shared/consts';
               <header-clear-storage></header-clear-storage>
             </div>
             <div nz-menu-item routerLink="/setting/tags">
-              <i nz-icon nzType="tag"></i>
+              <i nz-icon nzType="tag" class="mr-sm"></i>
               <span style="margin-left: 4px">{{ 'menu.extras.tags' | i18n }}</span>
             </div>
             <div nz-menu-item>
@@ -65,7 +67,7 @@ import { CONSTS } from 'src/app/shared/consts';
           <nz-avatar class="alain-default__aside-user-avatar" [nzSrc]="avatar"></nz-avatar>
           <div class="alain-default__aside-user-info">
             <strong>{{ user.name }}</strong>
-            <p class="mb0">{{ user.email }}</p>
+            <p class="mb0">{{ role }}</p>
           </div>
         </div>
       </ng-template>
@@ -79,8 +81,7 @@ import { CONSTS } from 'src/app/shared/consts';
     <global-footer style="border-top: 1px solid #e5e5e5; min-height: 120px; margin:0;">
       <div style="margin-top: 30px">
         HertzBeat {{ version }}<br />
-        Copyright
-        <i nz-icon nzType="copyright"></i> 2022
+        Copyright &copy; 2021-{{ currentYear }}
         <a href="https://hertzbeat.com" target="_blank">hertzbeat.com</a>
         <br />
         Licensed under the Apache License, Version 2.0
@@ -101,13 +102,24 @@ export class LayoutBasicComponent {
     logoExpanded: `./assets/brand_white.svg`,
     logoCollapsed: `./assets/logo.svg`
   };
-  avatar: string = `./assets/logo.svg`;
+  avatar: string = `./assets/img/avatar.svg`;
   searchToggleStatus = false;
   showSettingDrawer = !environment.production;
-  version = CONSTS.VERSION;
+  version = CONSTANTS.VERSION;
+  currentYear = new Date().getFullYear();
   get user(): User {
     return this.settings.user;
   }
 
-  constructor(private settings: SettingsService) {}
+  get role(): string {
+    let userTmp = this.settings.user;
+    if (userTmp == undefined || userTmp.role == undefined) {
+      return this.i18nSvc.fanyi('app.role.admin');
+    } else {
+      let roles: string[] = JSON.parse(userTmp.role);
+      return roles.length > 0 ? roles[0] : '';
+    }
+  }
+
+  constructor(private settings: SettingsService, @Inject(ALAIN_I18N_TOKEN) private i18nSvc: I18NService) {}
 }
