@@ -1,10 +1,10 @@
 import { ChangeDetectionStrategy, Component, HostListener } from '@angular/core';
-import * as screenfull from 'screenfull';
+import screenfull from 'screenfull';
 
 @Component({
   selector: 'header-fullscreen',
   template: `
-    <i nz-icon [nzType]="status ? 'fullscreen-exit' : 'fullscreen'"></i>
+    <i nz-icon class="mr-sm" [nzType]="status ? 'fullscreen-exit' : 'fullscreen'"></i>
     {{ (status ? 'menu.fullscreen.exit' : 'menu.fullscreen') | i18n }}
   `,
   host: {
@@ -14,19 +14,16 @@ import * as screenfull from 'screenfull';
 })
 export class HeaderFullScreenComponent {
   status = false;
-  private get sf(): screenfull.Screenfull {
-    return screenfull as screenfull.Screenfull;
-  }
 
   @HostListener('window:resize')
   _resize(): void {
-    this.status = this.sf.isFullscreen;
+    this.status = screenfull.isFullscreen;
   }
 
   @HostListener('click')
   _click(): void {
-    if (this.sf.isEnabled) {
-      this.sf.toggle();
+    if (screenfull.isEnabled) {
+      screenfull.toggle();
     }
   }
 }
