@@ -8,14 +8,22 @@ import { BooleanInput, InputBoolean } from '@delon/util/decorator';
   selector: 'header-i18n',
   template: `
     <div *ngIf="showLangText" nz-dropdown [nzDropdownMenu]="langMenu" nzPlacement="bottomRight">
-      <i nz-icon nzType="global"></i>
+      <i nz-icon class="mr-sm" nzType="global"></i>
       {{ 'menu.lang' | i18n }}
       <i nz-icon nzType="down"></i>
     </div>
-    <i *ngIf="!showLangText" nz-dropdown [nzDropdownMenu]="langMenu" nzPlacement="bottomRight" nz-icon nzType="global"></i>
+    <i
+      style="font-size: 25px; color: #b421cc"
+      *ngIf="!showLangText"
+      nz-dropdown
+      [nzDropdownMenu]="langMenu"
+      nzPlacement="bottomRight"
+      nz-icon
+      nzType="global"
+    ></i>
     <nz-dropdown-menu #langMenu="nzDropdownMenu">
       <ul nz-menu>
-        <li nz-menu-item *ngFor="let item of langs" [nzSelected]="item.code === curLangCode" (click)="change(item.code)">
+        <li nz-menu-item class="mr-sm" *ngFor="let item of langs" [nzSelected]="item.code === curLangCode" (click)="change(item.code)">
           <span role="img" [attr.aria-label]="item.text" class="pr-xs">{{ item.abbr }}</span>
           {{ item.text }}
         </li>
