@@ -1,10 +1,11 @@
 import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Inject, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { I18NService } from '@core';
-import { NoticeIconSelect, NoticeItem } from '@delon/abc/notice-icon';
+import { NoticeItem } from '@delon/abc/notice-icon';
 import { ALAIN_I18N_TOKEN } from '@delon/theme';
 import { NzI18nService } from 'ng-zorro-antd/i18n';
 import { NzMessageService } from 'ng-zorro-antd/message';
+import { finalize } from 'rxjs/operators';
 
 import { AlertService } from '../../../service/alert.service';
 
@@ -29,7 +30,7 @@ export class HeaderNotifyComponent implements OnInit {
       title: this.i18nSvc.fanyi('dashboard.alerts.title-no'),
       list: [],
       emptyText: this.i18nSvc.fanyi('dashboard.alerts.no'),
-      emptyImage: 'https://gw.alipayobjects.com/zos/rmsportal/wAhyIChODzsoKIOBHcBk.svg',
+      emptyImage: '/assets/img/empty-full.svg',
       clearText: this.i18nSvc.fanyi('dashboard.alerts.enter')
     }
   ];
@@ -54,48 +55,51 @@ export class HeaderNotifyComponent implements OnInit {
       return;
     }
     this.loading = true;
-    let loadAlerts$ = this.alertSvc.loadAlerts(0, undefined, undefined, 0, 5).subscribe(
-      message => {
-        loadAlerts$.unsubscribe();
-        if (message.code === 0) {
-          let page = message.data;
-          let alerts = page.content;
-          if (alerts == undefined) {
-            this.loading = false;
-            return;
+    let loadAlerts$ = this.alertSvc
+      .loadAlerts(0, undefined, undefined, 0, 5)
+      .pipe(
+        finalize(() => {
+          this.loading = false;
+        })
+      )
+      .subscribe(
+        message => {
+          loadAlerts$.unsubscribe();
+          if (message.code === 0) {
+            let page = message.data;
+            let alerts = page.content;
+            if (alerts == undefined) {
+              this.loading = false;
+              return;
+            }
+            this.data[0].list = [];
+            alerts.forEach(alert => {
+              let item = {
+                id: alert.id,
+                avatar: '/assets/img/notification.svg',
+                title: `${alert.tags?.monitorName}--${this.i18nSvc.fanyi(`alert.priority.${alert.priority}`)}`,
+                datetime: new Date(alert.lastAlarmTime).toLocaleString(),
+                color: 'blue',
+                type: this.i18nSvc.fanyi('dashboard.alerts.title-no')
+              };
+              this.data[0].list.push(item);
+            });
+            this.count = page.totalElements;
+          } else {
+            console.warn(message.msg);
           }
-          this.data[0].list = [];
-          alerts.forEach(alert => {
-            let item = {
-              id: alert.id,
-              avatar: 'https://gw.alipayobjects.com/zos/rmsportal/ThXAXghbEsBCCSDihZxY.png',
-              title: `${alert.tags?.monitorName}--${this.i18nSvc.fanyi(`alert.priority.${alert.priority}`)}`,
-              datetime: new Date(alert.lastTriggerTime).toLocaleString(),
-              color: 'blue',
-              type: this.i18nSvc.fanyi('dashboard.alerts.title-no')
-            };
-            this.data[0].list.push(item);
-          });
-          this.count = page.totalElements;
-        } else {
-          console.warn(message.msg);
+          this.loading = false;
+          this.cdr.detectChanges();
+        },
+        error => {
+          loadAlerts$.unsubscribe();
+          console.error(error);
+          this.loading = false;
         }
-        this.loading = false;
-        this.cdr.detectChanges();
-      },
-      error => {
-        loadAlerts$.unsubscribe();
-        console.error(error);
-        this.loading = false;
-      }
-    );
+      );
   }
 
   gotoAlertCenter(type: string): void {
     this.router.navigateByUrl(`/alert/center`);
-  }
-
-  select(res: NoticeIconSelect): void {
-    this.msg.success(`点击了 ${res.title} 的 ${res.item.title}`);
   }
 }
