@@ -5,8 +5,8 @@ export class Monitor {
   name!: string;
   app!: string;
   host!: string;
-  intervals: number = 120;
-  // 监控状态 0:未监控,1:可用,2:不可用,3:不可达,4:挂起
+  intervals: number = 60;
+  // 任务状态 0:未监控,1:可用,2:不可用
   status!: number;
   description!: string;
   creator!: string;
