@@ -2,14 +2,19 @@ import { ChangeDetectionStrategy, ChangeDetectorRef, Component, Inject, OnDestro
 import { Router } from '@angular/router';
 import { I18NService } from '@core';
 import { ALAIN_I18N_TOKEN } from '@delon/theme';
+import { CloudData } from 'angular-tag-cloud-module';
 import { EChartsOption } from 'echarts';
 import { NzMessageService } from 'ng-zorro-antd/message';
 import { fromEvent } from 'rxjs';
 
 import { Alert } from '../../pojo/Alert';
 import { AppCount } from '../../pojo/AppCount';
+import { CollectorSummary } from '../../pojo/CollectorSummary';
 import { AlertService } from '../../service/alert.service';
+import { CollectorService } from '../../service/collector.service';
 import { MonitorService } from '../../service/monitor.service';
+import { TagService } from '../../service/tag.service';
+import { formatTagName } from '../../shared/utils/common-util';
 
 @Component({
   selector: 'app-dashboard',
@@ -22,10 +27,70 @@ export class DashboardComponent implements OnInit, OnDestroy {
     private msg: NzMessageService,
     private monitorSvc: MonitorService,
     private alertSvc: AlertService,
+    private tagSvc: TagService,
+    private collectorSvc: CollectorService,
     @Inject(ALAIN_I18N_TOKEN) private i18nSvc: I18NService,
     private router: Router,
     private cdr: ChangeDetectorRef
   ) {}
+
+  // Tag Word Cloud
+  wordCloudData: CloudData[] = [];
+  defaultWordCloudData: CloudData[] = [
+    { text: 'HertzBeat', weight: 5 },
+    { text: 'Env:Prod', weight: 8 },
+    { text: 'Mysql', weight: 7 },
+    { text: 'Ping', weight: 5 },
+    { text: 'Product', weight: 6 },
+    { text: 'Env:Test', weight: 4 },
+    { text: 'TanCloud', weight: 3 },
+    { text: 'Bigdata', weight: 6 },
+    { text: 'Prod', weight: 4 },
+    { text: 'Dev', weight: 5 },
+    { text: 'Cache', weight: 7 },
+    { text: 'Region:US', weight: 7 },
+    { text: 'Region:CN', weight: 8 },
+    { text: 'Region:UK', weight: 7 },
+    { text: 'Localhost', weight: 6 },
+    { text: 'Cloud', weight: 8 },
+    { text: 'Network', weight: 5 },
+    { text: 'Custom', weight: 6 }
+  ];
+
+  refreshWordCloudContent(): void {
+    let tagsInit$ = this.tagSvc.loadTags(undefined, 1, 0, 10000).subscribe(
+      message => {
+        if (message.code === 0) {
+          let page = message.data;
+          let tags = page.content;
+          if (tags != null && tags.length != 0) {
+            let tmpData: CloudData[] = [];
+            tags.forEach(item => {
+              tmpData.push({
+                text: formatTagName(item),
+                weight: Math.random() * (10 - 5) + 5
+              });
+            });
+            this.wordCloudData = tmpData;
+          } else {
+            this.wordCloudData = this.defaultWordCloudData;
+          }
+          this.cdr.detectChanges();
+        } else {
+          console.warn(message.msg);
+        }
+        tagsInit$.unsubscribe();
+      },
+      error => {
+        tagsInit$.unsubscribe();
+        console.error(error.msg);
+      }
+    );
+  }
+
+  onTagCloudClick(data: CloudData): void {
+    this.router.navigate(['/monitors'], { queryParams: { tag: data.text } });
+  }
 
   // start 大类别数量信息
   appCountService: AppCount = new AppCount();
@@ -33,9 +98,44 @@ export class DashboardComponent implements OnInit, OnDestroy {
   appCountDb: AppCount = new AppCount();
   appCountMid: AppCount = new AppCount();
   appCountCustom: AppCount = new AppCount();
+  appCountProgram: AppCount = new AppCount();
+  appCountCache: AppCount = new AppCount();
+  appCountBigdata: AppCount = new AppCount();
+  appCountWebserver: AppCount = new AppCount();
+  appCountCn: AppCount = new AppCount();
+  appCountNetwork: AppCount = new AppCount();
+
+  slideConfig = {
+    infinite: true,
+    speed: 1200,
+    slidesToShow: 4,
+    slidesToScroll: 1,
+    autoplay: true,
+    autoplaySpeed: 1200,
+    rows: 1,
+    responsive: [
+      {
+        breakpoint: 1024,
+        settings: {
+          slidesToShow: 2.75,
+          slidesToScroll: 3,
+          speed: 4000,
+          infinite: true
+        }
+      },
+      {
+        breakpoint: 480,
+        settings: {
+          speed: 4000,
+          slidesToShow: 0.75,
+          slidesToScroll: 1
+        }
+      }
+    ]
+  };
 
   // start 数量全局概览
-  interval$!: number;
+  interval$!: any;
   appsCountLoading: boolean = true;
   appsCountTableData: any[] = [];
   appsCountEChartOption!: EChartsOption;
@@ -43,15 +143,36 @@ export class DashboardComponent implements OnInit, OnDestroy {
   appsCountEchartsInstance!: any;
   pageResize$!: any;
 
+  // collector list
+  collectors!: CollectorSummary[];
+
   // 告警列表
   alerts!: Alert[];
 
   ngOnInit(): void {
     this.appsCountTheme = {
       title: {
-        text: this.i18nSvc.fanyi('dashboard.monitors.title'),
-        subtext: this.i18nSvc.fanyi('dashboard.monitors.sub-title'),
-        left: 'center'
+        text: `{a|${this.i18nSvc.fanyi('dashboard.monitors.title')}}`,
+        subtext: `{b|${this.i18nSvc.fanyi('dashboard.monitors.sub-title')}}`,
+        left: 'center',
+        textStyle: {
+          rich: {
+            a: {
+              fontWeight: 'bolder',
+              align: 'center',
+              fontSize: 26
+            }
+          }
+        },
+        subtextStyle: {
+          rich: {
+            b: {
+              fontWeight: 'normal',
+              align: 'center',
+              fontSize: 14
+            }
+          }
+        }
       },
       tooltip: {
         trigger: 'item',
@@ -70,14 +191,15 @@ export class DashboardComponent implements OnInit, OnDestroy {
           name: this.i18nSvc.fanyi('dashboard.monitors.total'),
           type: 'pie',
           selectedMode: 'single',
-          color: '#722ED1',
+          color: '#3f51b5',
           radius: [0, '30%'],
           label: {
             position: 'center',
-            fontSize: 15,
+            fontSize: 38,
+            fontWeight: 'bolder',
             color: '#ffffff',
-            fontStyle: 'oblique',
-            formatter: '{a}:{c}'
+            fontStyle: 'normal',
+            formatter: '{c}'
           },
           labelLine: {
             show: false
@@ -92,7 +214,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
             length: 30
           },
           label: {
-            formatter: '{a|{a}}{abg|}\n{hr|}\n  {b|{b}：}{c}  {per|{d}%}  ',
+            formatter: '{a|{a}}{abg|}\n{hr|}\n  {b|{b}}   {per|{d}%}  ',
             backgroundColor: '#F6F8FC',
             borderColor: '#8C8D8E',
             borderWidth: 1,
@@ -128,8 +250,17 @@ export class DashboardComponent implements OnInit, OnDestroy {
     };
     this.alertsTheme = {
       title: {
-        subtext: this.i18nSvc.fanyi('dashboard.alerts.distribute'),
-        left: 'center'
+        subtext: `{b|${this.i18nSvc.fanyi('dashboard.alerts.distribute')}}`,
+        left: 'center',
+        subtextStyle: {
+          rich: {
+            b: {
+              fontWeight: 'bold',
+              align: 'center',
+              fontSize: 14
+            }
+          }
+        }
       },
       tooltip: {
         trigger: 'axis',
@@ -177,8 +308,17 @@ export class DashboardComponent implements OnInit, OnDestroy {
     };
     this.alertsDealTheme = {
       title: {
-        subtext: this.i18nSvc.fanyi('dashboard.alerts.deal'),
-        left: 'center'
+        subtext: `{b|${this.i18nSvc.fanyi('dashboard.alerts.deal')}}`,
+        left: 'center',
+        subtextStyle: {
+          rich: {
+            b: {
+              fontWeight: 'bold',
+              align: 'center',
+              fontSize: 14
+            }
+          }
+        }
       },
       tooltip: {
         formatter: '{b} : {c}%'
@@ -207,7 +347,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
     this.alertsLoading = true;
     this.refresh();
     // https://stackoverflow.com/questions/43908009/why-is-setinterval-in-an-angular-service-only-firing-one-time
-    this.interval$ = setInterval(this.refresh.bind(this), 30000);
+    this.interval$ = setInterval(this.refresh.bind(this), 80000);
     this.pageResize$ = fromEvent(window, 'resize').subscribe(event => {
       this.resizeChart();
     });
@@ -221,16 +361,24 @@ export class DashboardComponent implements OnInit, OnDestroy {
   }
 
   refresh(): void {
+    this.refreshWordCloudContent();
     this.refreshAppsCount();
     this.refreshAlertContentList();
+    this.refreshCollectorContentList();
     this.refreshAlertSummary();
   }
   refreshAppsCount(): void {
-    this.appCountService = new AppCount();
-    this.appCountOs = new AppCount();
-    this.appCountDb = new AppCount();
-    this.appCountMid = new AppCount();
-    this.appCountCustom = new AppCount();
+    let appCountService: AppCount = new AppCount();
+    let appCountOs: AppCount = new AppCount();
+    let appCountDb: AppCount = new AppCount();
+    let appCountMid: AppCount = new AppCount();
+    let appCountCustom: AppCount = new AppCount();
+    let appCountProgram: AppCount = new AppCount();
+    let appCountCache: AppCount = new AppCount();
+    let appCountWebserver: AppCount = new AppCount();
+    let appCountBigdata: AppCount = new AppCount();
+    let appCountCn: AppCount = new AppCount();
+    let appCountNetwork: AppCount = new AppCount();
     let dashboard$ = this.monitorSvc.getAppsMonitorSummary().subscribe(
       message => {
         dashboard$.unsubscribe();
@@ -251,42 +399,84 @@ export class DashboardComponent implements OnInit, OnDestroy {
             total = total + (app.size ? app.size : 0);
             switch (app.category) {
               case 'service':
-                this.appCountService.size += app.size;
-                this.appCountService.availableSize += app.availableSize;
-                this.appCountService.unAvailableSize += app.unAvailableSize;
-                this.appCountService.unManageSize += app.unManageSize;
-                this.appCountService.unReachableSize += app.unReachableSize;
+                appCountService.size += app.size;
+                appCountService.availableSize += app.availableSize;
+                appCountService.unAvailableSize += app.unAvailableSize;
+                appCountService.unManageSize += app.unManageSize;
                 break;
               case 'db':
-                this.appCountDb.size += app.size;
-                this.appCountDb.availableSize += app.availableSize;
-                this.appCountDb.unAvailableSize += app.unAvailableSize;
-                this.appCountDb.unManageSize += app.unManageSize;
-                this.appCountDb.unReachableSize += app.unReachableSize;
+                appCountDb.size += app.size;
+                appCountDb.availableSize += app.availableSize;
+                appCountDb.unAvailableSize += app.unAvailableSize;
+                appCountDb.unManageSize += app.unManageSize;
                 break;
               case 'os':
-                this.appCountOs.size += app.size;
-                this.appCountOs.availableSize += app.availableSize;
-                this.appCountOs.unAvailableSize += app.unAvailableSize;
-                this.appCountOs.unManageSize += app.unManageSize;
-                this.appCountOs.unReachableSize += app.unReachableSize;
+                appCountOs.size += app.size;
+                appCountOs.availableSize += app.availableSize;
+                appCountOs.unAvailableSize += app.unAvailableSize;
+                appCountOs.unManageSize += app.unManageSize;
                 break;
               case 'mid':
-                this.appCountMid.size += app.size;
-                this.appCountMid.availableSize += app.availableSize;
-                this.appCountMid.unAvailableSize += app.unAvailableSize;
-                this.appCountMid.unManageSize += app.unManageSize;
-                this.appCountMid.unReachableSize += app.unReachableSize;
+                appCountMid.size += app.size;
+                appCountMid.availableSize += app.availableSize;
+                appCountMid.unAvailableSize += app.unAvailableSize;
+                appCountMid.unManageSize += app.unManageSize;
                 break;
               case 'custom':
-                this.appCountCustom.size += app.size;
-                this.appCountCustom.availableSize += app.availableSize;
-                this.appCountCustom.unAvailableSize += app.unAvailableSize;
-                this.appCountCustom.unManageSize += app.unManageSize;
-                this.appCountCustom.unReachableSize += app.unReachableSize;
+                appCountCustom.size += app.size;
+                appCountCustom.availableSize += app.availableSize;
+                appCountCustom.unAvailableSize += app.unAvailableSize;
+                appCountCustom.unManageSize += app.unManageSize;
+                break;
+              case 'program':
+                appCountProgram.size += app.size;
+                appCountProgram.availableSize += app.availableSize;
+                appCountProgram.unAvailableSize += app.unAvailableSize;
+                appCountProgram.unManageSize += app.unManageSize;
+                break;
+              case 'bigdata':
+                appCountBigdata.size += app.size;
+                appCountBigdata.availableSize += app.availableSize;
+                appCountBigdata.unAvailableSize += app.unAvailableSize;
+                appCountBigdata.unManageSize += app.unManageSize;
+                break;
+              case 'webserver':
+                appCountWebserver.size += app.size;
+                appCountWebserver.availableSize += app.availableSize;
+                appCountWebserver.unAvailableSize += app.unAvailableSize;
+                appCountWebserver.unManageSize += app.unManageSize;
+                break;
+              case 'cache':
+                appCountCache.size += app.size;
+                appCountCache.availableSize += app.availableSize;
+                appCountCache.unAvailableSize += app.unAvailableSize;
+                appCountCache.unManageSize += app.unManageSize;
+                break;
+              case 'cn':
+                appCountCn.size += app.size;
+                appCountCn.availableSize += app.availableSize;
+                appCountCn.unAvailableSize += app.unAvailableSize;
+                appCountCn.unManageSize += app.unManageSize;
+                break;
+              case 'network':
+                appCountNetwork.size += app.size;
+                appCountNetwork.availableSize += app.availableSize;
+                appCountNetwork.unAvailableSize += app.unAvailableSize;
+                appCountNetwork.unManageSize += app.unManageSize;
                 break;
             }
           });
+          this.appCountService = appCountService;
+          this.appCountOs = appCountOs;
+          this.appCountDb = appCountDb;
+          this.appCountMid = appCountMid;
+          this.appCountCustom = appCountCustom;
+          this.appCountBigdata = appCountBigdata;
+          this.appCountCache = appCountCache;
+          this.appCountProgram = appCountProgram;
+          this.appCountWebserver = appCountWebserver;
+          this.appCountNetwork = appCountNetwork;
+          this.appCountCn = appCountCn;
           // @ts-ignore
           this.appsCountTheme.series[0].data = [{ value: total, name: this.i18nSvc.fanyi('dashboard.monitors.total') }];
           // @ts-ignore
@@ -343,12 +533,6 @@ export class DashboardComponent implements OnInit, OnDestroy {
   alertsEchartsInstance!: any;
   alertsLoading: boolean = true;
 
-  refreshAlerts(): void {
-    this.alertsEChartOption = this.alertsTheme;
-    this.cdr.detectChanges();
-    this.alertsLoading = false;
-  }
-
   // start 告警处理率
   alertsDealEChartOption!: EChartsOption;
   alertsDealTheme!: EChartsOption;
@@ -356,7 +540,7 @@ export class DashboardComponent implements OnInit, OnDestroy {
   alertsDealLoading: boolean = true;
 
   refreshAlertContentList(): void {
-    let alertsInit$ = this.alertSvc.loadAlerts(undefined, undefined, undefined, 0, 4).subscribe(
+    let alertsInit$ = this.alertSvc.loadAlerts(undefined, undefined, undefined, 0, 10).subscribe(
       message => {
         if (message.code === 0) {
           let page = message.data;
@@ -369,6 +553,24 @@ export class DashboardComponent implements OnInit, OnDestroy {
       },
       error => {
         alertsInit$.unsubscribe();
+        console.error(error.msg);
+      }
+    );
+  }
+
+  refreshCollectorContentList(): void {
+    let collectorInit$ = this.collectorSvc.getCollectors().subscribe(
+      message => {
+        if (message.code === 0) {
+          this.collectors = message.data.content;
+          this.cdr.detectChanges();
+        } else {
+          console.warn(message.msg);
+        }
+        collectorInit$.unsubscribe();
+      },
+      error => {
+        collectorInit$.unsubscribe();
         console.error(error.msg);
       }
     );
@@ -417,6 +619,8 @@ export class DashboardComponent implements OnInit, OnDestroy {
           console.warn(message.msg);
         }
         alertSummaryInit$.unsubscribe();
+        this.alertsLoading = false;
+        this.alertsDealLoading = false;
       },
       error => {
         alertSummaryInit$.unsubscribe();
