@@ -1,4 +1,5 @@
 import { Component, Input } from '@angular/core';
+import { NzNotificationService } from 'ng-zorro-antd/notification';
 
 import { MonitorService } from '../../../service/monitor.service';
 
@@ -19,6 +20,8 @@ export class MonitorDataTableComponent {
   }
   private _monitorId!: number;
   @Input()
+  app!: string;
+  @Input()
   metrics!: string;
 
   time!: any;
@@ -27,7 +30,7 @@ export class MonitorDataTableComponent {
   rowValues!: any[];
   isTable: boolean = true;
 
-  constructor(private monitorSvc: MonitorService) {}
+  constructor(private monitorSvc: MonitorService, private notifySvc: NzNotificationService) {}
 
   loadData() {
     // 读取实时指标数据
@@ -42,7 +45,8 @@ export class MonitorDataTableComponent {
             this.isTable = false;
             this.rowValues = this.valueRows[0].values;
           }
-        } else {
+        } else if (message.code !== 0) {
+          this.notifySvc.warning(`${this.metrics}:${message.msg}`, '');
           console.info(`${this.metrics}:${message.msg}`);
         }
       },
