@@ -9,6 +9,7 @@ import { NzRadioModule } from 'ng-zorro-antd/radio';
 import { NzSpaceModule } from 'ng-zorro-antd/space';
 import { NzSwitchModule } from 'ng-zorro-antd/switch';
 import { NzTagModule } from 'ng-zorro-antd/tag';
+import { NzUploadModule } from 'ng-zorro-antd/upload';
 import { NgxEchartsModule } from 'ngx-echarts';
 
 import { MonitorDataChartComponent } from './monitor-data-chart/monitor-data-chart.component';
@@ -41,7 +42,8 @@ const COMPONENTS: Array<Type<void>> = [
     NzLayoutModule,
     NzSpaceModule,
     NzCollapseModule,
-    ClipboardModule
+    ClipboardModule,
+    NzUploadModule
   ],
   declarations: COMPONENTS
 })
