@@ -1,4 +1,4 @@
-import { HttpClient, HttpParams } from '@angular/common/http';
+import { HttpClient, HttpParams, HttpResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 
@@ -9,6 +9,7 @@ import { Page } from '../pojo/Page';
 
 const alert_define_uri = '/alert/define';
 const alert_defines_uri = '/alert/defines';
+const export_defines_uri = '/alert/defines/export';
 
 @Injectable({
   providedIn: 'root'
@@ -47,7 +48,7 @@ export class AlertDefineService {
     return this.http.delete<Message<any>>(alert_defines_uri, options);
   }
 
-  public getAlertDefines(pageIndex: number, pageSize: number): Observable<Message<Page<AlertDefine>>> {
+  public getAlertDefines(search: string | undefined, pageIndex: number, pageSize: number): Observable<Message<Page<AlertDefine>>> {
     pageIndex = pageIndex ? pageIndex : 0;
     pageSize = pageSize ? pageSize : 8;
     // 注意HttpParams是不可变对象 需要保存set后返回的对象为最新对象
@@ -58,7 +59,25 @@ export class AlertDefineService {
       pageIndex: pageIndex,
       pageSize: pageSize
     });
+    if (search != undefined && search.trim() != '') {
+      httpParams = httpParams.append('search', search.trim());
+    }
     const options = { params: httpParams };
     return this.http.get<Message<Page<AlertDefine>>>(alert_defines_uri, options);
+  }
+
+  public exportAlertDefines(defineIds: Set<number>, type: string): Observable<HttpResponse<Blob>> {
+    let httpParams = new HttpParams();
+    defineIds.forEach(defineId => {
+      // 注意HttpParams是不可变对象 需要保存append后返回的对象为最新对象
+      // append方法可以叠加同一key, set方法会把key之前的值覆盖只留一个key-value
+      httpParams = httpParams.append('ids', defineId);
+    });
+    httpParams = httpParams.append('type', type);
+    return this.http.get(export_defines_uri, {
+      params: httpParams,
+      observe: 'response',
+      responseType: 'blob'
+    });
   }
 }
