@@ -5,22 +5,28 @@ import { RouterModule } from '@angular/router';
 import { DelonACLModule } from '@delon/acl';
 import { DelonFormModule } from '@delon/form';
 import { AlainThemeModule } from '@delon/theme';
+import { NzBreadCrumbModule } from 'ng-zorro-antd/breadcrumb';
+import { NzTagModule } from 'ng-zorro-antd/tag';
 
+import { HelpMassageShowComponent } from './components/help-massage-show/help-massage-show.component';
 import { KeyValueInputComponent } from './components/key-value-input/key-value-input.component';
+import { MetricsFieldInputComponent } from './components/metrics-field-input/metrics-field-input.component';
+import { ElapsedTimePipe } from './pipe/elapsed-time.pipe';
+import { I18nElsePipe } from './pipe/i18n-else.pipe';
 import { TimezonePipe } from './pipe/timezone.pipe';
 import { SHARED_DELON_MODULES } from './shared-delon.module';
 import { SHARED_ZORRO_MODULES } from './shared-zorro.module';
 
 // #region third libs
 
-const THIRDMODULES: Array<Type<void>> = [];
+const ThirdModules: Array<Type<void>> = [];
 
 // #endregion
 
 // #region your components & directives
 
-const COMPONENTS: Array<Type<void>> = [KeyValueInputComponent];
-const DIRECTIVES: Array<Type<void>> = [TimezonePipe];
+const COMPONENTS: Array<Type<void>> = [KeyValueInputComponent, HelpMassageShowComponent, MetricsFieldInputComponent];
+const DIRECTIVES: Array<Type<void>> = [TimezonePipe, I18nElsePipe, ElapsedTimePipe];
 
 // #endregion
 
@@ -36,12 +42,15 @@ const DIRECTIVES: Array<Type<void>> = [TimezonePipe];
     ...SHARED_DELON_MODULES,
     ...SHARED_ZORRO_MODULES,
     // third libs
-    ...THIRDMODULES
+    ...ThirdModules,
+    NzBreadCrumbModule,
+    NzTagModule
   ],
   declarations: [
     // your components
     ...COMPONENTS,
-    ...DIRECTIVES
+    ...DIRECTIVES,
+    HelpMassageShowComponent
   ],
   exports: [
     CommonModule,
@@ -54,7 +63,7 @@ const DIRECTIVES: Array<Type<void>> = [TimezonePipe];
     ...SHARED_DELON_MODULES,
     ...SHARED_ZORRO_MODULES,
     // third libs
-    ...THIRDMODULES,
+    ...ThirdModules,
     // your components
     ...COMPONENTS,
     ...DIRECTIVES
