@@ -2,10 +2,7 @@
 // `ng build ---prod` replaces `environment.ts` with `environment.prod.ts`.
 // The list of file replacements can be found in `angular.json`.
 
-import { DelonMockModule } from '@delon/mock';
 import { Environment } from '@delon/theme';
-
-import * as MOCK_DATA from '../../_mock';
 
 export const environment = {
   production: false,
@@ -13,8 +10,7 @@ export const environment = {
   api: {
     baseUrl: 'http://localhost:1157/api/',
     refreshTokenEnabled: true
-  },
-  modules: [DelonMockModule.forRoot({ data: MOCK_DATA })]
+  }
 } as Environment;
 
 /*
